requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
torch>=2.0.0+cu118
torchaudio>=2.0.0+cu118
pyannote.audio>=3.1.0
huggingface-hub>=0.16.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
librosa>=0.10.0 